fixes with settings screen(add settings applying inside recycler but state saving in db is the rest of task)

+++ app/src/main/java/com/example/noteslist/presentation/model/SettingsUiState.kt
@@ -1,6 +1,6 @@
 package com.example.noteslist.presentation.model
 
 data class SettingsUiState(
-    val stackSpacing: Int,
+    val stackSpacing: Float,
     val stackMaxVisible: Int
 )
